requirements.txt
python-telegram-bot==12.8
urllib3==1.26.6